+++ socialhome/authapi/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class AuthApiConfig(AppConfig):
-    name = "socialhome.authapi"
-    verbose_name = "SPA Authentication API"