+++ socialhome/streams/enums.py
@@ -1,30 +1,20 @@
 from django.utils.translation import gettext_lazy as _
-from enumfields import Enum
+from enum_properties import StrEnumProperties
 
 
-class StreamType(Enum):
-    CONTENT = "content"
-    CUSTOM = "custom"
-    FOLLOWED = "followed"
-    LIMITED = "limited"
-    LOCAL = "local"
-    PROFILE_ALL = "profile_all"
-    PROFILE_PINNED = "profile_pinned"
-    PUBLIC = "public"
-    TAG = "tag"
-    TAGS = "tags"
+class StreamType(StrEnumProperties):
+    label: str
 
-    class Labels:
-        CONTENT = _("Content")
-        CUSTOM = _("Custom")
-        FOLLOWED = _("Followed")
-        LIMITED = _("Limited")
-        LOCAL = _("Local")
-        PROFILE_ALL = _("Profile (all)")
-        PROFILE_PINNED = _("Profile (pinned)")
-        PUBLIC = _("Public")
-        TAG = _("Tag")
-        TAGS = _("Tags")
+    CONTENT = "content", _("Content")
+    CUSTOM = "custom", _("Custom")
+    FOLLOWED = "followed", _("Followed")
+    LIMITED = "limited", _("Limited")
+    LOCAL = "local", _("Local")
+    PROFILE_ALL = "profile_all", _("Profile (all)")
+    PROFILE_PINNED = "profile_pinned", _("Profile (pinned)")
+    PUBLIC = "public", _("Public")
+    TAG = "tag", _("Tag")
+    TAGS = "tags", _("Tags")
 
     @classmethod
     def to_dict(cls):