+++ socialhome/streams/streams.py
@@ -186,7 +186,7 @@ class BaseStream:
     key_base = ["sh", "streams"]
     notify_for_shares = True
     ordering = "-created"
-    paginate_by = 15
+    paginate_by = 30
     redis = None
     stream_type = None
     unfetched_content = False
@@ -390,7 +390,7 @@ class BaseStream:
 
     def should_cache_stream(self, cls, user):
         return cls in CACHED_STREAM_CLASSES and settings.SOCIALHOME_STREAMS_PRECACHE_SIZE
-        
+
 
 class FollowedStream(BaseStream):
     stream_type = StreamType.FOLLOWED
@@ -523,9 +523,9 @@ class TagsStream(BaseStream):
 
 CACHED_STREAM_CLASSES = (
     FollowedStream,
-    ProfileAllStream,
     LimitedStream,
     LocalStream,
+    ProfileAllStream,
     PublicStream,
     TagStream,
     TagsStream,