+++ socialhome/users/preferences.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 from dynamic_preferences.types import Section, ChoicePreference, BooleanPreference
 from dynamic_preferences.users.registries import user_preferences_registry
@@ -22,12 +23,3 @@ class LandingPage(ChoicePreference):
     default = "followed"
     verbose_name = _("Landing page")
     help_text = _("Choose which page you want to see as the landing page.")
-
-
-@user_preferences_registry.register
-class UseNewUI(BooleanPreference):
-    section = generic
-    name = "use_new_ui"
-    default = True
-    verbose_name = _("Use new UI")
-    help_text = _("Try the new SPA UI.")