+++ socialhome/users/tests/factories.py
@@ -37,6 +37,10 @@ class UserFactory(factory.django.DjangoModelFactory):
             verified=True,
         )
 
+    @classmethod
+    def _after_postgeneration(cls, obj, create, results):
+        obj.save()
+
 
 class LimitedUserFactory(UserFactory):
     @classmethod
@@ -133,27 +137,31 @@ class ProfileFactory(factory.django.DjangoModelFactory):
 
 
     @factory.post_generation
-    def set_fid(self, extracted, created, **kwargs):
+    def set_fid(self, create, extracted, **kwargs):
         if self.fid == "placeholder":
             self.fid = f"{settings.SOCIALHOME_URL}/u/{self.user.username}/" \
                 if self.user else f"{settings.SOCIALHOME_URL}/p/{self.uuid}/"
 
     # noinspection PyAttributeOutsideInit
     @factory.post_generation
-    def set_handle(self, extracted, created, **kwargs):
-        if extracted is False or self.handle:
+    def set_handle(self, create, extracted, **kwargs):
+        if self.handle:
             return
 
         # Set handle sometimes, sometimes not, but also allow passing in True to force
-        if extracted is True or randint(0, 100) > 50:
+        if create is True or extracted is True or randint(0, 100) > 50:
             self.handle = self.finger
 
     @factory.post_generation
-    def with_key(self, extracted, created, **kwargs):
+    def with_key(self, create, extracted, **kwargs):
         if not extracted:
             return
 
         self.generate_new_rsa_key(bits=1024)
+
+    @classmethod
+    def _after_postgeneration(cls, obj, create, results):
+        obj.save()
 
 
 class PublicProfileFactory(ProfileFactory):